adding (int64_t) to allow compilation on 64 bit machines ticket:539
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@6161 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ util/tf/test/tf_unittest_future.cpp
@@ -17,7 +17,7 @@ void seed_rand()
 
 TEST(tf, NoExtrapolationExceptionFromParent)
 {
-  tf::Transformer mTR(true, ros::Duration(1000000LL), ros::Duration(0LL));
+  tf::Transformer mTR(true, ros::Duration((int64_t)1000000LL), ros::Duration((int64_t)0LL));
   
 
 
@@ -54,7 +54,7 @@ TEST(tf, NoExtrapolationExceptionFromParent)
 
 TEST(tf, ExtrapolationFromOneValue)
 {
-  tf::Transformer mTR(true, ros::Duration(1000000LL), ros::Duration(0LL));
+  tf::Transformer mTR(true, ros::Duration((int64_t)1000000LL), ros::Duration((int64_t)0LL));
